+++ src/library/Uno.Cupertino/Styles/Controls/PasswordBox.xaml
@@ -63,7 +63,7 @@
 												Value="1" />
 									</VisualState.Setters>
 									<VisualState.StateTriggers>
-										<StateTrigger IsActive="{Binding Password, Converter={StaticResource CupertinoEmptyToTrue}, RelativeSource={RelativeSource TemplatedParent}}" />
+										<StateTrigger IsActive="{Binding Password, Converter={StaticResource EmptyToTrueConverter}, RelativeSource={RelativeSource TemplatedParent}}" />
 									</VisualState.StateTriggers>
 								</VisualState>
 							</VisualStateGroup>
