+++ src/library/Uno.Material/MaterialColorsV2.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using Uno.Themes;
 using Windows.Foundation.Metadata;
 
 #if WinUI
