+++ src/library/Uno.Material/MaterialTheme.cs
@@ -1,6 +1,8 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using Uno.Themes;
+
 
 #if WinUI
 using Microsoft.UI;
