+++ src/library/Uno.Material/Styles/Controls/v1/Button.xaml
@@ -49,7 +49,7 @@
 				Value="True" />
 
 		<!-- Workaround for WinUI issue: https://github.com/microsoft/microsoft-ui-xaml/issues/6388 -->
-		<Setter Property="um:ControlExtensions.Icon"
+		<Setter Property="ut:ControlExtensions.Icon"
 				Value="{x:Null}" />
 
 		<Setter Property="Template">
@@ -150,7 +150,7 @@
 									</Grid.ColumnDefinitions>
 
 									<ContentPresenter x:Name="IconPresenter"
-													  Content="{Binding Path=(um:ControlExtensions.Icon), RelativeSource={RelativeSource TemplatedParent}}"
+													  Content="{Binding Path=(ut:ControlExtensions.Icon), RelativeSource={RelativeSource TemplatedParent}}"
 													  Margin="0,0,8,0"
 													  MaxHeight="34"
 													  MaxWidth="34"
@@ -158,7 +158,7 @@
 													  HorizontalAlignment="Stretch"
 													  VerticalAlignment="Stretch"
 													  VerticalContentAlignment="Center"
-													  Visibility="{Binding Path=(um:ControlExtensions.Icon), RelativeSource={RelativeSource TemplatedParent}, Converter={StaticResource NullToCollapsedConverterConverter}, FallbackValue=Collapsed, TargetNullValue=Collapsed}" />
+													  Visibility="{Binding Path=(ut:ControlExtensions.Icon), RelativeSource={RelativeSource TemplatedParent}, Converter={StaticResource NullToCollapsedConverter}, FallbackValue=Collapsed, TargetNullValue=Collapsed}" />
 
 									<ContentPresenter Grid.Column="1"
 													  Content="{Binding Content, RelativeSource={RelativeSource Mode=TemplatedParent}, Converter={StaticResource ToUpperConverter}}"
@@ -216,7 +216,7 @@
 				Value="True" />
 
 		<!-- Workaround for WinUI issue: https://github.com/microsoft/microsoft-ui-xaml/issues/6388 -->
-		<Setter Property="um:ControlExtensions.Icon"
+		<Setter Property="ut:ControlExtensions.Icon"
 				Value="{x:Null}" />
 
 		<Setter Property="Template">
@@ -323,7 +323,7 @@
 									</Grid.ColumnDefinitions>
 
 									<ContentPresenter x:Name="IconPresenter"
-													  Content="{Binding Path=(um:ControlExtensions.Icon), RelativeSource={RelativeSource TemplatedParent}}"
+													  Content="{Binding Path=(ut:ControlExtensions.Icon), RelativeSource={RelativeSource TemplatedParent}}"
 													  Margin="0,0,8,0"
 													  MaxHeight="34"
 													  MaxWidth="34"
@@ -331,7 +331,7 @@
 													  HorizontalAlignment="Stretch"
 													  VerticalAlignment="Stretch"
 													  VerticalContentAlignment="Center"
-													  Visibility="{Binding Path=(um:ControlExtensions.Icon), RelativeSource={RelativeSource TemplatedParent}, Converter={StaticResource NullToCollapsedConverterConverter}, FallbackValue=Collapsed, TargetNullValue=Collapsed}" />
+													  Visibility="{Binding Path=(ut:ControlExtensions.Icon), RelativeSource={RelativeSource TemplatedParent}, Converter={StaticResource NullToCollapsedConverter}, FallbackValue=Collapsed, TargetNullValue=Collapsed}" />
 
 									<ContentPresenter Grid.Column="1"
 													  Content="{Binding Content, RelativeSource={RelativeSource Mode=TemplatedParent}, Converter={StaticResource ToUpperConverter}}"
@@ -384,7 +384,7 @@
 				Value="True" />
 
 		<!-- Workaround for WinUI issue: https://github.com/microsoft/microsoft-ui-xaml/issues/6388 -->
-		<Setter Property="um:ControlExtensions.Icon"
+		<Setter Property="ut:ControlExtensions.Icon"
 				Value="{x:Null}" />
 
 		<Setter Property="Template">
@@ -488,7 +488,7 @@
 									</Grid.ColumnDefinitions>
 
 									<ContentPresenter x:Name="IconPresenter"
-													  Content="{Binding Path=(um:ControlExtensions.Icon), RelativeSource={RelativeSource TemplatedParent}}"
+													  Content="{Binding Path=(ut:ControlExtensions.Icon), RelativeSource={RelativeSource TemplatedParent}}"
 													  Margin="0,0,8,0"
 													  MaxHeight="34"
 													  MaxWidth="34"
@@ -496,7 +496,7 @@
 													  HorizontalAlignment="Stretch"
 													  VerticalAlignment="Stretch"
 													  VerticalContentAlignment="Center"
-													  Visibility="{Binding Path=(um:ControlExtensions.Icon), RelativeSource={RelativeSource TemplatedParent}, Converter={StaticResource NullToCollapsedConverterConverter}, FallbackValue=Collapsed, TargetNullValue=Collapsed}" />
+													  Visibility="{Binding Path=(ut:ControlExtensions.Icon), RelativeSource={RelativeSource TemplatedParent}, Converter={StaticResource NullToCollapsedConverter}, FallbackValue=Collapsed, TargetNullValue=Collapsed}" />
 
 									<ContentPresenter Grid.Column="1"
 													  Content="{Binding Content, RelativeSource={RelativeSource Mode=TemplatedParent}, Converter={StaticResource ToUpperConverter}}"
