+++ src/library/Uno.Material/Styles/Controls/v1/TextBox.xaml
@@ -100,7 +100,7 @@
 				Value="50" />
 
 		<!-- Workaround for WinUI issue: https://github.com/microsoft/microsoft-ui-xaml/issues/6388 -->
-		<Setter Property="um:ControlExtensions.Icon"
+		<Setter Property="ut:ControlExtensions.Icon"
 				Value="{x:Null}" />
 
 		<Setter Property="Template">
@@ -207,8 +207,8 @@
 
 							<Border Margin="{TemplateBinding Padding}">
 								<ContentPresenter x:Name="IconPresenter"
-												  Visibility="{Binding Path=(um:ControlExtensions.Icon), RelativeSource={RelativeSource TemplatedParent}, Converter={StaticResource NullToCollapsedConverterConverter}, FallbackValue=Collapsed, TargetNullValue=Collapsed}"
-												  Content="{Binding Path=(um:ControlExtensions.Icon), RelativeSource={RelativeSource TemplatedParent}}"
+												  Visibility="{Binding Path=(ut:ControlExtensions.Icon), RelativeSource={RelativeSource TemplatedParent}, Converter={StaticResource NullToCollapsedConverter}, FallbackValue=Collapsed, TargetNullValue=Collapsed}"
+												  Content="{Binding Path=(ut:ControlExtensions.Icon), RelativeSource={RelativeSource TemplatedParent}}"
 												  MaxHeight="34"
 												  MaxWidth="34"
 												  MinWidth="25"
@@ -314,7 +314,7 @@
 				Value="50" />
 
 		<!-- Workaround for WinUI issue: https://github.com/microsoft/microsoft-ui-xaml/issues/6388 -->
-		<Setter Property="um:ControlExtensions.Icon"
+		<Setter Property="ut:ControlExtensions.Icon"
 				Value="{x:Null}" />
 
 		<Setter Property="Template">
@@ -418,8 +418,8 @@
 
 							<Border Margin="{TemplateBinding Padding}">
 								<ContentPresenter x:Name="IconPresenter"
-                                                  Visibility="{Binding Path=(um:ControlExtensions.Icon), RelativeSource={RelativeSource TemplatedParent}, Converter={StaticResource NullToCollapsedConverterConverter}, FallbackValue=Collapsed, TargetNullValue=Collapsed}"
-												  Content="{Binding Path=(um:ControlExtensions.Icon), RelativeSource={RelativeSource TemplatedParent}}"
+                                                  Visibility="{Binding Path=(ut:ControlExtensions.Icon), RelativeSource={RelativeSource TemplatedParent}, Converter={StaticResource NullToCollapsedConverter}, FallbackValue=Collapsed, TargetNullValue=Collapsed}"
+												  Content="{Binding Path=(ut:ControlExtensions.Icon), RelativeSource={RelativeSource TemplatedParent}}"
 												  MaxHeight="34"
 												  MaxWidth="34"
 												  MinWidth="25"
