+++ src/library/Uno.Material/Styles/Controls/v1/ToggleButton.xaml
@@ -4,6 +4,7 @@
 					xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
 					xmlns:win="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
 					xmlns:um="using:Uno.Material"
+					xmlns:ut="using:Uno.Themes"
 					xmlns:xamarin="http://uno.ui/xamarin"
 					mc:Ignorable="d xamarin">
 
@@ -186,7 +187,7 @@
 				Value="Center" />
 
 		<!-- Workaround for WinUI issue: https://github.com/microsoft/microsoft-ui-xaml/issues/6388 -->
-		<Setter Property="um:ControlExtensions.AlternateContent"
+		<Setter Property="ut:ControlExtensions.AlternateContent"
 				Value="{x:Null}" />
 		
 		<Setter Property="Template">
@@ -294,7 +295,7 @@
 						<!-- Alternate Content -->
 						<ContentPresenter x:Name="AlternateContentPresenter"
 										  AutomationProperties.AccessibilityView="Raw"
-										  Content="{Binding Path=(um:ControlExtensions.AlternateContent), RelativeSource={RelativeSource TemplatedParent}}"
+										  Content="{Binding Path=(ut:ControlExtensions.AlternateContent), RelativeSource={RelativeSource TemplatedParent}}"
 										  HorizontalContentAlignment="{TemplateBinding HorizontalContentAlignment}"
 										  VerticalContentAlignment="{TemplateBinding VerticalContentAlignment}"
 										  Visibility="Collapsed" />
