+++ src/samples/UWP/Uno.Themes.Samples.Shared/Content/Controls/ComboBoxSamplePage.xaml
@@ -7,6 +7,7 @@
 	  xmlns:android="http://uno.ui/android"
 	  xmlns:ios="http://uno.ui/ios"
 	  xmlns:um="using:Uno.Material"
+	  xmlns:ut="using:Uno.Themes"
 	  xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
 	  xmlns:smtx="using:ShowMeTheXAML"
 	  mc:Ignorable="d android ios">
@@ -130,9 +131,9 @@
 									<ComboBox Style="{StaticResource MaterialComboBoxStyle}"
 											  PlaceholderText="Placeholder"
 											  ItemsSource="abcdefghijklmn">
-										<um:ControlExtensions.Icon>
+										<ut:ControlExtensions.Icon>
 											<PathIcon Data="{StaticResource Icon_Mail}" />
-										</um:ControlExtensions.Icon>
+										</ut:ControlExtensions.Icon>
 									</ComboBox>
 								</smtx:XamlDisplay>
 
@@ -152,9 +153,9 @@
 											  PlaceholderText="Placeholder"
 											  ItemsSource="abcdefghijklmn"
 											  IsEnabled="False">
-										<um:ControlExtensions.Icon>
+										<ut:ControlExtensions.Icon>
 											<PathIcon Data="{StaticResource Icon_Mail}" />
-										</um:ControlExtensions.Icon>
+										</ut:ControlExtensions.Icon>
 									</ComboBox>
 								</smtx:XamlDisplay>
 							</StackPanel>
