+++ src/samples/UWP/Uno.Themes.Samples.Shared/Content/Controls/FlyoutSamplePage.xaml
@@ -3,7 +3,7 @@
 	  xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
 	  xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
 	  xmlns:smtx="using:ShowMeTheXAML"
-	  xmlns:um="using:Uno.Material"
+	  xmlns:ut="using:Uno.Themes"
 	  xmlns:sample="using:Uno.Themes.Samples"
 	  xmlns:toolkit="using:Uno.UI.Toolkit">
 	<Page.Resources>
@@ -126,9 +126,9 @@
 													HorizontalAlignment="Stretch"
 													HorizontalContentAlignment="Left"
 													Style="{StaticResource MaterialTextButtonStyle}">
-												<um:ControlExtensions.Icon>
+												<ut:ControlExtensions.Icon>
 													<SymbolIcon Symbol="Share" />
-												</um:ControlExtensions.Icon>
+												</ut:ControlExtensions.Icon>
 											</Button>
 
 											<Button Grid.Row="1"
@@ -136,9 +136,9 @@
 													HorizontalAlignment="Stretch"
 													HorizontalContentAlignment="Left"
 													Style="{StaticResource MaterialTextButtonStyle}">
-												<um:ControlExtensions.Icon>
+												<ut:ControlExtensions.Icon>
 													<SymbolIcon Symbol="Link" />
-												</um:ControlExtensions.Icon>
+												</ut:ControlExtensions.Icon>
 											</Button>
 
 											<Button Grid.Row="2"
@@ -146,9 +146,9 @@
 													HorizontalAlignment="Stretch"
 													HorizontalContentAlignment="Left"
 													Style="{StaticResource MaterialTextButtonStyle}">
-												<um:ControlExtensions.Icon>
+												<ut:ControlExtensions.Icon>
 													<SymbolIcon Symbol="Edit" />
-												</um:ControlExtensions.Icon>
+												</ut:ControlExtensions.Icon>
 											</Button>
 
 											<Button Grid.Row="3"
@@ -156,9 +156,9 @@
 													HorizontalAlignment="Stretch"
 													HorizontalContentAlignment="Left"
 													Style="{StaticResource MaterialTextButtonStyle}">
-												<um:ControlExtensions.Icon>
+												<ut:ControlExtensions.Icon>
 													<SymbolIcon Symbol="Delete" />
-												</um:ControlExtensions.Icon>
+												</ut:ControlExtensions.Icon>
 											</Button>
 										</Grid>
 									</Flyout>
@@ -425,9 +425,9 @@
 													HorizontalAlignment="Stretch"
 													HorizontalContentAlignment="Left"
 													Style="{StaticResource MaterialTextButtonStyle}">
-												<um:ControlExtensions.Icon>
+												<ut:ControlExtensions.Icon>
 													<SymbolIcon Symbol="Share" />
-												</um:ControlExtensions.Icon>
+												</ut:ControlExtensions.Icon>
 											</Button>
 
 											<Button Grid.Row="1"
@@ -435,9 +435,9 @@
 													HorizontalAlignment="Stretch"
 													HorizontalContentAlignment="Left"
 													Style="{StaticResource MaterialTextButtonStyle}">
-												<um:ControlExtensions.Icon>
+												<ut:ControlExtensions.Icon>
 													<SymbolIcon Symbol="Link" />
-												</um:ControlExtensions.Icon>
+												</ut:ControlExtensions.Icon>
 											</Button>
 
 											<Button Grid.Row="2"
@@ -445,9 +445,9 @@
 													HorizontalAlignment="Stretch"
 													HorizontalContentAlignment="Left"
 													Style="{StaticResource MaterialTextButtonStyle}">
-												<um:ControlExtensions.Icon>
+												<ut:ControlExtensions.Icon>
 													<SymbolIcon Symbol="Edit" />
-												</um:ControlExtensions.Icon>
+												</ut:ControlExtensions.Icon>
 											</Button>
 
 											<Button Grid.Row="3"
@@ -455,9 +455,9 @@
 													HorizontalAlignment="Stretch"
 													HorizontalContentAlignment="Left"
 													Style="{StaticResource MaterialTextButtonStyle}">
-												<um:ControlExtensions.Icon>
+												<ut:ControlExtensions.Icon>
 													<SymbolIcon Symbol="Delete" />
-												</um:ControlExtensions.Icon>
+												</ut:ControlExtensions.Icon>
 											</Button>
 										</Grid>
 									</Flyout>
