+++ src/samples/UWP/Uno.Themes.Samples.Shared/Content/Controls/PasswordBoxSamplePage.xaml
@@ -8,6 +8,7 @@
 	  xmlns:ios="http://uno.ui/ios"
 	  xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
 	  xmlns:um="using:Uno.Material"
+	  xmlns:ut="using:Uno.Themes"
 	  xmlns:smtx="using:ShowMeTheXAML"
 	  mc:Ignorable="d android ios">
 
@@ -61,9 +62,9 @@
 						<smtx:XamlDisplay UniqueKey="Material_PasswordBoxSamplePage_6"
 										  Style="{StaticResource XamlDisplayBelowStyle}">
 							<PasswordBox Style="{StaticResource MaterialFilledPasswordBoxStyle}">
-								<um:ControlExtensions.Icon>
+								<ut:ControlExtensions.Icon>
 									<SymbolIcon Symbol="Favorite" />
-								</um:ControlExtensions.Icon>
+								</ut:ControlExtensions.Icon>
 							</PasswordBox>
 						</smtx:XamlDisplay>
 
@@ -72,9 +73,9 @@
 										  Style="{StaticResource XamlDisplayBelowStyle}">
 							<PasswordBox PlaceholderText="Filled with Icon"
 										 Style="{StaticResource MaterialFilledPasswordBoxStyle}">
-								<um:ControlExtensions.Icon>
+								<ut:ControlExtensions.Icon>
 									<SymbolIcon Symbol="Favorite" />
-								</um:ControlExtensions.Icon>
+								</ut:ControlExtensions.Icon>
 							</PasswordBox>
 						</smtx:XamlDisplay>
 
@@ -83,9 +84,9 @@
 										  Style="{StaticResource XamlDisplayBelowStyle}">
 							<PasswordBox PlaceholderText="Filled with Icon"
 										 Style="{StaticResource MaterialOutlinedPasswordBoxStyle}">
-								<um:ControlExtensions.Icon>
+								<ut:ControlExtensions.Icon>
 									<SymbolIcon Symbol="Favorite" />
-								</um:ControlExtensions.Icon>
+								</ut:ControlExtensions.Icon>
 							</PasswordBox>
 						</smtx:XamlDisplay>
 					</StackPanel>
