+++ src/samples/UWP/Uno.Themes.Samples.Shared/Content/Controls/TextBoxSamplePage.xaml
@@ -6,6 +6,7 @@
 	  xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
 	  xmlns:smtx="using:ShowMeTheXAML"
 	  xmlns:um="using:Uno.Material"
+	  xmlns:ut="using:Uno.Themes"
 	  mc:Ignorable="d">
 
 	<Grid Background="{ThemeResource ApplicationPageBackgroundThemeBrush}">
@@ -72,9 +73,9 @@
 						<smtx:XamlDisplay UniqueKey="Material_TextBoxSamplePage_8"
 										  Style="{StaticResource XamlDisplayBelowStyle}">
 							<TextBox Style="{StaticResource MaterialFilledTextBoxStyle}">
-								<um:ControlExtensions.Icon>
+								<ut:ControlExtensions.Icon>
 									<SymbolIcon Symbol="Favorite" />
-								</um:ControlExtensions.Icon>
+								</ut:ControlExtensions.Icon>
 							</TextBox>
 						</smtx:XamlDisplay>
 
@@ -83,9 +84,9 @@
 										  Style="{StaticResource XamlDisplayBelowStyle}">
 							<TextBox PlaceholderText="Filled with Icon"
 									 Style="{StaticResource MaterialFilledTextBoxStyle}">
-								<um:ControlExtensions.Icon>
+								<ut:ControlExtensions.Icon>
 									<SymbolIcon Symbol="Favorite" />
-								</um:ControlExtensions.Icon>
+								</ut:ControlExtensions.Icon>
 							</TextBox>
 						</smtx:XamlDisplay>
 
@@ -94,9 +95,9 @@
 										  Style="{StaticResource XamlDisplayBelowStyle}">
 							<TextBox PlaceholderText="Outlined with Icon"
 									 Style="{StaticResource MaterialOutlinedTextBoxStyle}">
-								<um:ControlExtensions.Icon>
+								<ut:ControlExtensions.Icon>
 									<SymbolIcon Symbol="Favorite" />
-								</um:ControlExtensions.Icon>
+								</ut:ControlExtensions.Icon>
 							</TextBox>
 						</smtx:XamlDisplay>
 					</StackPanel>
@@ -177,9 +178,9 @@
 									 Style="{StaticResource MaterialFilledTextBoxStyle}"
 									 TextWrapping="Wrap"
 									 AcceptsReturn="True">
-								<um:ControlExtensions.Icon>
+								<ut:ControlExtensions.Icon>
 									<SymbolIcon Symbol="Favorite" />
-								</um:ControlExtensions.Icon>
+								</ut:ControlExtensions.Icon>
 							</TextBox>
 						</smtx:XamlDisplay>
 
@@ -191,9 +192,9 @@
 									 Style="{StaticResource MaterialOutlinedTextBoxStyle}"
 									 TextWrapping="Wrap"
 									 AcceptsReturn="True">
-								<um:ControlExtensions.Icon>
+								<ut:ControlExtensions.Icon>
 									<SymbolIcon Symbol="Favorite" />
-								</um:ControlExtensions.Icon>
+								</ut:ControlExtensions.Icon>
 							</TextBox>
 						</smtx:XamlDisplay>
 
@@ -202,9 +203,9 @@
 										  Style="{StaticResource XamlDisplayBelowStyle}">
 							<TextBox PlaceholderText="Filled with Icon"
 									 Style="{StaticResource MaterialFilledTextBoxStyle}">
-								<um:ControlExtensions.Icon>
+								<ut:ControlExtensions.Icon>
 									<SymbolIcon Symbol="Favorite" />
-								</um:ControlExtensions.Icon>
+								</ut:ControlExtensions.Icon>
 							</TextBox>
 						</smtx:XamlDisplay>
 
@@ -213,9 +214,9 @@
 										  Style="{StaticResource XamlDisplayBelowStyle}">
 							<TextBox PlaceholderText="Outlined with Icon"
 									 Style="{StaticResource MaterialOutlinedTextBoxStyle}">
-								<um:ControlExtensions.Icon>
+								<ut:ControlExtensions.Icon>
 									<SymbolIcon Symbol="Favorite" />
-								</um:ControlExtensions.Icon>
+								</ut:ControlExtensions.Icon>
 							</TextBox>
 						</smtx:XamlDisplay>
 					</StackPanel>
