+++ src/samples/UWP/Uno.Themes.Samples.Shared/Content/Controls/ToggleButtonSamplePage.xaml
@@ -6,6 +6,7 @@
 	  xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
 	  xmlns:sample="using:Uno.Themes.Samples"
 	  xmlns:smtx="using:ShowMeTheXAML"
+	  xmlns:ut="using:Uno.Themes"
 	  xmlns:um="using:Uno.Material"
 	  mc:Ignorable="d">
 
@@ -21,9 +22,9 @@
 								<ToggleButton.Content>
 									<FontIcon Glyph="&#xE160;" />
 								</ToggleButton.Content>
-								<um:ControlExtensions.AlternateContent>
+								<ut:ControlExtensions.AlternateContent>
 									<FontIcon Glyph="&#xE132;" />
-								</um:ControlExtensions.AlternateContent>
+								</ut:ControlExtensions.AlternateContent>
 							</ToggleButton>
 						</smtx:XamlDisplay>
 
@@ -35,9 +36,9 @@
 								<ToggleButton.Content>
 									<FontIcon Glyph="&#xE160;" />
 								</ToggleButton.Content>
-								<um:ControlExtensions.AlternateContent>
+								<ut:ControlExtensions.AlternateContent>
 									<FontIcon Glyph="&#xE132;" />
-								</um:ControlExtensions.AlternateContent>
+								</ut:ControlExtensions.AlternateContent>
 							</ToggleButton>
 						</smtx:XamlDisplay>
 
@@ -49,9 +50,9 @@
 								<ToggleButton.Content>
 									<FontIcon Glyph="&#xE160;" />
 								</ToggleButton.Content>
-								<um:ControlExtensions.AlternateContent>
+								<ut:ControlExtensions.AlternateContent>
 									<FontIcon Glyph="&#xE132;" />
-								</um:ControlExtensions.AlternateContent>
+								</ut:ControlExtensions.AlternateContent>
 							</ToggleButton>
 						</smtx:XamlDisplay>
 
@@ -64,9 +65,9 @@
 								<ToggleButton.Content>
 									<FontIcon Glyph="&#xE160;" />
 								</ToggleButton.Content>
-								<um:ControlExtensions.AlternateContent>
+								<ut:ControlExtensions.AlternateContent>
 									<FontIcon Glyph="&#xE132;" />
-								</um:ControlExtensions.AlternateContent>
+								</ut:ControlExtensions.AlternateContent>
 							</ToggleButton>
 						</smtx:XamlDisplay>
 					</StackPanel>
