+++ src/samples/UWP/Uno.Themes.Samples.Shared/Content/Extensions/ControlExtensionsSamplePage.xaml
@@ -6,6 +6,7 @@
 	  xmlns:ios="http://uno.ui/ios"
 	  xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
 	  xmlns:um="using:Uno.Material"
+	  xmlns:ut="using:Uno.Themes"
 	  xmlns:smtx="using:ShowMeTheXAML"
 	  mc:Ignorable="d android ios">
 	<Page.Resources>
@@ -86,10 +87,10 @@
 								  Style="{StaticResource XamlDisplayBelowStyle}">
 					<TextBox PlaceholderText="Filled"
 							 Style="{StaticResource MaterialFilledTextBoxStyle}">
-						<um:ControlExtensions.Icon>
+						<ut:ControlExtensions.Icon>
 							<SymbolIcon Symbol="SolidStar"
 										Foreground="{StaticResource MaterialPrimaryBrush}" />
-						</um:ControlExtensions.Icon>
+						</ut:ControlExtensions.Icon>
 					</TextBox>
 				</smtx:XamlDisplay>
 
@@ -101,10 +102,10 @@
 					<ComboBox Style="{StaticResource MaterialComboBoxStyle}"
 							  PlaceholderText="Placeholder"
 							  ItemsSource="abcdefg">
-						<um:ControlExtensions.Icon>
+						<ut:ControlExtensions.Icon>
 							<SymbolIcon Symbol="SolidStar"
 										Foreground="{StaticResource MaterialPrimaryBrush}" />
-						</um:ControlExtensions.Icon>
+						</ut:ControlExtensions.Icon>
 					</ComboBox>
 				</smtx:XamlDisplay>
 
@@ -121,7 +122,7 @@
 				<smtx:XamlDisplay UniqueKey="ControlExtensionsSamplePage_ToggleButtonText"
 								  Style="{StaticResource XamlDisplayBelowStyle}">
 					<ToggleButton Content="On"
-								  um:ControlExtensions.AlternateContent="Off"
+								  ut:ControlExtensions.AlternateContent="Off"
 								  Style="{StaticResource MaterialTextToggleButtonStyle}" />
 				</smtx:XamlDisplay>
 
@@ -131,9 +132,9 @@
 						<ToggleButton.Content>
 							<PathIcon Data="{StaticResource Icon_more_horizontal}" />
 						</ToggleButton.Content>
-						<um:ControlExtensions.AlternateContent>
+						<ut:ControlExtensions.AlternateContent>
 							<PathIcon Data="{StaticResource Icon_more_vertical}" />
-						</um:ControlExtensions.AlternateContent>
+						</ut:ControlExtensions.AlternateContent>
 					</ToggleButton>
 				</smtx:XamlDisplay>
