+++ src/samples/UWP/Uno.Themes.Samples.Shared/Content/NestedSamples/M3MaterialCommandBarSample_NestedPage1.xaml.cs
@@ -12,6 +12,7 @@ using Windows.UI.Xaml.Data;
 using Windows.UI.Xaml.Input;
 using Windows.UI.Xaml.Media;
 using Windows.UI.Xaml.Navigation;
+using Uno.Themes;
 
 namespace Uno.Themes.Samples.Content.NestedSamples
 {
