+++ src/samples/UWP/Uno.Themes.Samples.Shared/Content/OverviewPage.xaml
@@ -3,6 +3,7 @@
 	  xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
 	  xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
 	  xmlns:um="using:Uno.Material"
+	  xmlns:ut="using:Uno.Themes"
 	  xmlns:local="using:Uno.Themes.Samples.Content"
 	  xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
 	  xmlns:sample="using:Uno.Themes.Samples"
@@ -57,9 +58,9 @@
 							<TextBox AutomationProperties.AutomationId="MaterialFilledIconTextBox"
 									 PlaceholderText="Filled with icon"
 									 Style="{StaticResource MaterialFilledTextBoxStyle}">
-								<um:ControlExtensions.Icon>
+								<ut:ControlExtensions.Icon>
 									<SymbolIcon Symbol="Favorite" />
-								</um:ControlExtensions.Icon>
+								</ut:ControlExtensions.Icon>
 							</TextBox>
 
 							<TextBox AutomationProperties.AutomationId="MaterialOutlinedTextBox"
@@ -69,9 +70,9 @@
 							<TextBox AutomationProperties.AutomationId="MaterialOutlinedWithIconTextBox"
 									 PlaceholderText="Outlined with icon"
 									 Style="{StaticResource MaterialOutlinedTextBoxStyle}">
-								<um:ControlExtensions.Icon>
+								<ut:ControlExtensions.Icon>
 									<SymbolIcon Symbol="Favorite" />
-								</um:ControlExtensions.Icon>
+								</ut:ControlExtensions.Icon>
 							</TextBox>
 						</StackPanel>
 					</sample:OverviewSampleView>
@@ -164,9 +165,9 @@
 							<TextBox AutomationProperties.AutomationId="M3MaterialFilledIconTextBox"
 									 PlaceholderText="Filled with icon"
 									 Style="{StaticResource MaterialFilledTextBoxStyle}">
-								<um:ControlExtensions.Icon>
+								<ut:ControlExtensions.Icon>
 									<SymbolIcon Symbol="Favorite" />
-								</um:ControlExtensions.Icon>
+								</ut:ControlExtensions.Icon>
 							</TextBox>
 
 							<TextBox AutomationProperties.AutomationId="M3MaterialOutlinedTextBox"
@@ -176,9 +177,9 @@
 							<TextBox AutomationProperties.AutomationId="M3MaterialOutlinedWithIconTextBox"
 									 PlaceholderText="Outlined with icon"
 									 Style="{StaticResource MaterialOutlinedTextBoxStyle}">
-								<um:ControlExtensions.Icon>
+								<ut:ControlExtensions.Icon>
 									<SymbolIcon Symbol="Favorite" />
-								</um:ControlExtensions.Icon>
+								</ut:ControlExtensions.Icon>
 							</TextBox>
 						</StackPanel>
 					</sample:OverviewSampleView>
